+++ Makefile
@@ -12,22 +12,22 @@ clean-client:
 	rm -f ./client
 
 run-master:
-	go run app/services/master/main.go
+	go run cmd/master/main.go
 
 run-client:
-	go run app/services/client/main.go
+	go run cmd/client/main.go
 
 run-chunkserver:
-	go run app/services/chunkserver/main.go
+	go run cmd/chunkserver/main.go
 
 build-master:
-	go build -o master app/services/master/main.go
+	go build -o master cmd/master/main.go
 
 build-chunkserver:
-	go build -o chunkserver app/services/chunkserver/main.go
+	go build -o chunkserver cmd/chunkserver/main.go
 
 build-client:
-	go build -o chunkserver app/services/chunkserver/main.go
+	go build -o client cmd/client/main.go
 
 tidy:
 	go mod tidy
