+++ core/master/master.go
@@ -2,21 +2,21 @@ package master
 
 import (
 	"errors"
+	"github.com/pyropy/dfs/core/chunk_metadata_service"
+	"github.com/pyropy/dfs/core/constants"
+	filemeta "github.com/pyropy/dfs/core/file_metadata_service"
+	chunkServerRPC "github.com/pyropy/dfs/rpc/chunkserver"
 	"log"
 	"math/rand"
 	"net/rpc"
 
 	"github.com/google/uuid"
-	chunkmeta "github.com/pyropy/dfs/business/core/chunk_metadata_service"
-	"github.com/pyropy/dfs/business/core/constants"
-	filemeta "github.com/pyropy/dfs/business/core/file_metadata_service"
-	chunkServerRPC "github.com/pyropy/dfs/business/rpc/chunkserver"
 )
 
 type Master struct {
 	*LeaseService
 	*filemeta.FileMetadataService
-	*chunkmeta.ChunkMetadataService
+	*chunkmetadataservice.ChunkMetadataService
 	*ChunkServerMetadataService
 }
 
@@ -30,7 +30,7 @@ func NewMaster() *Master {
 	return &Master{
 		LeaseService:               NewLeaseService(),
 		FileMetadataService:        filemeta.NewFileMetadataService(),
-		ChunkMetadataService:       chunkmeta.NewChunkMetadataService(),
+		ChunkMetadataService:       chunkmetadataservice.NewChunkMetadataService(),
 		ChunkServerMetadataService: NewChunkServerMetadataService(),
 	}
 }
@@ -39,7 +39,7 @@ func NewMaster() *Master {
 // CreateNewFile selects chunk servers and instructs them to create N number of chunks with predefined IDs
 func (m *Master) CreateNewFile(filePath string, fileSizeBytes, repFactor, chunkSizeBytes int) (*filemeta.FileMetadata, []uuid.UUID, error) {
 	var chunkIds []uuid.UUID
-	var chunkMetadata []chunkmeta.ChunkMetadata
+	var chunkMetadata []chunkmetadataservice.ChunkMetadata
 	var chunkServerIds []uuid.UUID
 
 	fileExists := m.FileMetadataService.CheckFileExists(filePath)
@@ -60,7 +60,7 @@ func (m *Master) CreateNewFile(filePath string, fileSizeBytes, repFactor, chunkS
 		chunkID := uuid.New()
 		chunkIds = append(chunkIds, chunkID)
 		fileMetadata.Chunks = append(fileMetadata.Chunks, chunkID)
-		chunk := chunkmeta.NewChunkMetadata(chunkID, i, chunkVersion, chunkServerIds)
+		chunk := chunkmetadataservice.NewChunkMetadata(chunkID, i, chunkVersion, chunkServerIds)
 		chunkMetadata = append(chunkMetadata, chunk)
 
 		for _, chunkServer := range chunkServers {
